requirements.txt
Flask
loguru
Werkzeug
python-multipart
opencv-python
Pillow
tqdm
pandas
transformers
timm